# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/staff/kb/articles": {
      "get": {
        "description": "Every status, drafts included, most recently changed first. Filter by status, category or words.",
        "operationId": "KbStaffController_articles",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "One or more statuses; all by default",
            "schema": {
              "minItems": 1,
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "draft",
                  "published",
                  "archived"
                ]
              }
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Words to look for",
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page, 1 to 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "`nextCursor` from the previous page",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffArticlePage_Output"
                }
              }
            }
          },
          "400": {
            "description": "A filter or the cursor is invalid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid session for this realm",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing the `kb:read` permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffSession": []
          }
        ],
        "summary": "List articles",
        "tags": [
          "Knowledge base (staff)"
        ]
      },
      "post": {
        "description": "The slug is made from the title unless you give one. Raw HTML and unsafe links are removed from the markdown before it is stored; the response shows what was kept.",
        "operationId": "KbStaffController_createArticle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffArticle_Output"
                }
              }
            }
          },
          "400": {
            "description": "The body is invalid (`validation-error`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid session for this realm",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing the `kb:write` permission\n\nThe Origin or the CSRF token was missing or wrong (`csrf-rejected`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The brand isn't one of yours",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The slug is taken in this brand (`already-exists`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The category isn't in this brand",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffSession": [],
            "csrfToken": []
          }
        ],
        "summary": "Write a draft",
        "tags": [
          "Knowledge base (staff)"
        ]
      }
    },
    "/api/v1/staff/kb/articles/{articleId}": {
      "get": {
        "operationId": "KbStaffController_article",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffArticle_Output"
                }
              }
            }
          },
          "401": {
            "description": "No valid session for this realm",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing the `kb:read` permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such article in your brands",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffSession": []
          }
        ],
        "summary": "One article, with its markdown",
        "tags": [
          "Knowledge base (staff)"
        ]
      },
      "patch": {
        "description": "Saving a published article publishes the new text as the next version straight away, so it also needs `kb:publish`. Drafts and archived articles stay as they are.",
        "operationId": "KbStaffController_updateArticle",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffArticle_Output"
                }
              }
            }
          },
          "400": {
            "description": "The body is invalid (`validation-error`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid session for this realm",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing `kb:write`, or `kb:publish` for a published article\n\nThe Origin or the CSRF token was missing or wrong (`csrf-rejected`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such article in your brands",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The slug is taken in this brand (`already-exists`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The category isn't in this brand",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffSession": [],
            "csrfToken": []
          }
        ],
        "summary": "Edit an article",
        "tags": [
          "Knowledge base (staff)"
        ]
      }
    },
    "/api/v1/staff/kb/articles/{articleId}/publish": {
      "post": {
        "description": "Puts a draft or archived article on the help centre as its next version. Already published is a no-op.",
        "operationId": "KbStaffController_publish",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffArticle_Output"
                }
              }
            }
          },
          "401": {
            "description": "No valid session for this realm",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing the `kb:publish` permission\n\nThe Origin or the CSRF token was missing or wrong (`csrf-rejected`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such article in your brands",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffSession": [],
            "csrfToken": []
          }
        ],
        "summary": "Publish",
        "tags": [
          "Knowledge base (staff)"
        ]
      }
    },
    "/api/v1/staff/kb/articles/{articleId}/unpublish": {
      "post": {
        "description": "Returns the article to draft: off the help centre and out of AI retrieval.",
        "operationId": "KbStaffController_unpublish",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffArticle_Output"
                }
              }
            }
          },
          "401": {
            "description": "No valid session for this realm",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing the `kb:publish` permission\n\nThe Origin or the CSRF token was missing or wrong (`csrf-rejected`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such article in your brands",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffSession": [],
            "csrfToken": []
          }
        ],
        "summary": "Unpublish",
        "tags": [
          "Knowledge base (staff)"
        ]
      }
    },
    "/api/v1/staff/kb/articles/{articleId}/archive": {
      "post": {
        "description": "Keeps the article for the record but takes it off the help centre and out of AI retrieval.",
        "operationId": "KbStaffController_archive",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffArticle_Output"
                }
              }
            }
          },
          "401": {
            "description": "No valid session for this realm",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing the `kb:publish` permission\n\nThe Origin or the CSRF token was missing or wrong (`csrf-rejected`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such article in your brands",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffSession": [],
            "csrfToken": []
          }
        ],
        "summary": "Archive",
        "tags": [
          "Knowledge base (staff)"
        ]
      }
    },
    "/api/v1/staff/kb/categories": {
      "get": {
        "description": "In display order.",
        "operationId": "KbStaffController_categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categories_Output"
                }
              }
            }
          },
          "401": {
            "description": "No valid session for this realm",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing the `kb:read` permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffSession": []
          }
        ],
        "summary": "List categories",
        "tags": [
          "Knowledge base (staff)"
        ]
      },
      "post": {
        "operationId": "KbStaffController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category_Output"
                }
              }
            }
          },
          "400": {
            "description": "The body is invalid (`validation-error`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid session for this realm",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing the `kb:write` permission\n\nThe Origin or the CSRF token was missing or wrong (`csrf-rejected`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The brand isn't one of yours",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The slug is taken in this brand (`already-exists`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffSession": [],
            "csrfToken": []
          }
        ],
        "summary": "Add a category",
        "tags": [
          "Knowledge base (staff)"
        ]
      }
    },
...
  },
  "info": {
    "title": "DSD Unified Customer Support API",
    "description": "The API behind the DSD customer app and the DSD agent app.\n\nAll routes except `/health` and `/ready` live under `/api/v1`. Errors use RFC 9457 problem details (`application/problem+json`); every response carries an `X-Request-Id` header that matches the `requestId` in an error body and the server logs.\n\n## Signing in\n\nCustomers and staff are separate realms with separate routes, cookies and sessions: `/api/v1/auth/customer/*` and `/api/v1/customer/*` for customers, `/api/v1/auth/staff/*` and `/api/v1/staff/*` for staff. A session from one realm is refused (401) by every route of the other.\n\nSigning in sets an HttpOnly session cookie and a readable CSRF cookie, and returns the CSRF token in the body too (it is also in `me`). Every POST, PUT, PATCH and DELETE must come from a trusted origin, and once signed in must send that token in the `X-CSRF-Token` header.\n\nTo try it here, sign in with \"Try it out\" on a `login` route: the browser keeps the cookies, and this page copies the CSRF token into the header for you.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "customerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-dsd_customer_session",
        "description": "Set by the customer login route. HttpOnly, so scripts can't read it."
      },
      "staffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-dsd_staff_session",
        "description": "Set by the staff login route. HttpOnly, so scripts can't read it."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "The csrfToken from login or me. Required with the session cookie on every state-changing request."
      }
    },
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Problem type URI; `about:blank` when the HTTP status says it all"
          },
          "title": {
            "type": "string",
            "description": "Short summary of the problem type"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "description": "Explanation specific to this occurrence",
            "type": "string"
          },
          "instance": {
            "description": "The request path that produced the problem",
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "description": "Matches the X-Request-Id response header and the server logs"
          },
          "errors": {
            "description": "Present on validation problems: one entry per invalid field",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "description": "Dot-separated path to the invalid field; empty for the whole body"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          },
          "allowedTransitions": {
            "description": "Present on `invalid-status-transition`: the statuses the ticket can move to now",
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "open",
                "pending_customer",
                "resolved",
                "closed"
              ]
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "requestId"
        ]
      }
    }
  }
}